+++ packages/tests/src/e2e/vs/BotHappyPathCommon.ts
